+++ apps/backend/src/Obscura.Domain/Media/AudioTrack.cs
@@ -33,7 +33,6 @@ public sealed class AudioTrack : Entity {
     protected override IEnumerable<EntityCapability> CreateDefaultCapabilities() =>
     [
         new CapabilityRating(),
-        new CapabilityImages(),
         new CapabilityLinks(),
         new CapabilityFlags(),
         new CapabilityFiles(),