+++ apps/backend/src/Obscura.Domain/Media/Book.cs
@@ -50,7 +50,6 @@ public sealed class Book : Entity {
     protected override IEnumerable<EntityCapability> CreateDefaultCapabilities() =>
     [
         new CapabilityRating(),
-        new CapabilityImages(),
         new CapabilityLinks(),
         new CapabilityFlags(),
         new CapabilityFiles(),