+++ apps/backend/src/Obscura.Domain/Media/Collection.cs
@@ -43,7 +43,6 @@ public sealed class Collection : Entity {
     protected override IEnumerable<EntityCapability> CreateDefaultCapabilities() =>
     [
         new CapabilityRating(),
-        new CapabilityImages(),
         new CapabilityLinks(),
         new CapabilityFlags(),
         new CapabilityFiles()