+++ apps/backend/src/Obscura.Domain/Media/Video.cs
@@ -24,7 +24,6 @@ public sealed class Video : Entity {
     protected override IEnumerable<EntityCapability> CreateDefaultCapabilities() =>
     [
         new CapabilityRating(),
-        new CapabilityImages(),
         new CapabilityLinks(),
         new CapabilityFlags(),
         new CapabilityFiles(),