+++ apps/backend/src/Obscura.Domain/Taxonomy/Tag.cs
@@ -18,7 +18,6 @@ public sealed class Tag : Entity {
     protected override IEnumerable<EntityCapability> CreateDefaultCapabilities() =>
     [
         new CapabilityRating(),
-        new CapabilityImages(),
         new CapabilityLinks(),
         new CapabilityFlags(),
         new CapabilityFiles()